feat: Abstract invitations handling for `useInvitedRoom`
